Restructured STT and enabled prosody tags for generative Polly
